+++ docs/changelog.md
@@ -1,4 +1,7 @@
 # TBD
+### Changes
+* Added TypeScript case to regenerate-protobuf-bindings.sh so that now it considers TypeScript within the shell script
+* Generated TypeScript bindings which can now be ouputted by the wrapper script which calls down to developer-tools
 
 # 0.2.0
 ### Changes